re-arrange outcome displays
